// frontend/src/app/api/generated/types.gen.ts
export type PageUserInfoDto = {
    totalPages?: number;
    totalElements?: number;
    size?: number;
    content?: Array<UserInfoDto>;
    number?: number;
    sort?: SortObject;
    pageable?: PageableObject;
    numberOfElements?: number;
    first?: boolean;
    last?: boolean;
    empty?: boolean;
};

export type PageableObject = {
    paged?: boolean;
    pageNumber?: number;
    pageSize?: number;
    sort?: SortObject;
    unpaged?: boolean;
    offset?: number;
};

export type SortObject = {
    unsorted?: boolean;
    sorted?: boolean;
    empty?: boolean;
};

export type UserInfoDto = {
    id?: number;
    email?: string;
    display_name?: string;
};